tools/rados/rados.cc: fix UNINTENDED_INTEGER_DIVISION
Fix for:

CID 1297882 (#1 of 1): Result is not floating-point (UNINTENDED_INTEGER_DIVISION)
 integer_division: Dividing integer expressions ns and 1000000000UL, and then
 converting the integer quotient to type float. Any remainder, or fractional
 part of the quotient, is ignored.

+++ src/tools/rados/rados.cc
@@ -569,7 +569,7 @@ public:
     utime_t now = ceph_clock_now(g_ceph_context);
     now -= start_time;
     uint64_t ns = now.nsec();
-    float total = ns / 1000000000;
+    float total = (float) ns / 1000000000.0;
     total += now.sec();
     return total;
   }
